% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimCMk.R, R/simcmk1.R
\name{SimCMk}
\alias{SimCMk}
\title{simulation de chaine de markov}
\usage{
SimCMk(P, mu, n)

SimCMk(P, mu, n)
}
\arguments{
\item{P}{matrix of transition}

\item{mu}{numeric vector of initial distrebution}

\item{n}{numeric number of repetition}
}
\description{
simulation de chaine de markov
}
